Extract SVG geometry conversion
